server/database/requirements-postgres.txt
psycopg2-binary>=2.9.9,<2.10
